update backend custom setting.fix admin log on post a field with array.add new method for backend article search.fix ueditor on 32 bit oerpation system int auto convert to float

+++ backend/web/static/js/plugins/ueditor/php/Uploader.class.php
@@ -308,7 +308,7 @@ class Uploader
         $format = str_replace("{filename}", $oriName, $format);
 
         //替换随机字符串
-        $randNum = rand(1, 10000000000) . rand(1, 10000000000);
+        $randNum = rand(1,  (int)10000000000) . rand(1,  (int)10000000000);
         if (preg_match("/\{rand\:([\d]*)\}/i", $format, $matches)) {
             $format = preg_replace("/\{rand\:[\d]*\}/i", substr($randNum, 0, $matches[1]), $format);
         }
